questões

+++ exercicios/007.py
@@ -1,4 +1,6 @@
-n1 = int(input('Digite a primeira nota: '))
-n2 = int(input('Digite a segunda nota: '))
+#Media entre dois números
+
+n1 = float(input('Digite a primeira nota: '))
+n2 = float(input('Digite a segunda nota: '))
 media = (n1 + n2) / 2
 print('A media é {:.1f}'.format(media))
